Support for arrays and in-expr
Add support for arrays and array literals of the form [ e1, e2, ...].
All elements in the array must have the same type. The empty array
literal [] is invalid.

Also add a new expression in the form of: <left> in <right>. An in-expr
always evaluates to boolean. <right> must be an array of the same type
as <left>.

E.g.

2 in [ 1, 2, 3]

evaluates to true.

in-expr currently only supports arrays of number or string.