fix: move global initialization to appease race detector
We enable `yyErrorVerbose` to get extra detailed syntax error messages,
but setting it in `Parse()` does not make Go's race detector happy when
the parser is invoked from multiple goroutines. It is a global after
all, being written to without synchronization. Moving the setting of the
variable to `init()` solves the problem.